Feat_[Client, Server]: finish adding review comments
