+++ client/src/components/reviews/ReviewEdit.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+type Props = {};
+
+export default function ReviewEdit({}: Props) {
+  return <div>ReviewEdit</div>;
+}
